requirements.txt
discord.py
PyGithub
flask